Add multiple method to pick the last tag
Previously the biggest tag (compared as semantic version) was chosen regarding
on what branch it was. Now the default behavior is to consider only tags on the
specified branch.

The previous default had the following limitations:

- you can't do for example a patch bump on an old version
- sometimes it makes sense to consider the latest tag applied not to compare
  them as semantic versions.

A new CLI option was added to select the search strategy for the last tag.